Restructure some of the components into seperate files
